+++ src/main/java/de/srendi/advancedperipherals/AdvancedPeripherals.java
@@ -1,17 +1,15 @@
 package de.srendi.advancedperipherals;
 
-import de.srendi.advancedperipherals.client.HudOverlayHandler;
 import de.srendi.advancedperipherals.common.addons.APAddons;
 import de.srendi.advancedperipherals.common.configuration.APConfig;
 import de.srendi.advancedperipherals.common.setup.Registration;
 import de.srendi.advancedperipherals.common.village.VillageStructures;
-import de.srendi.advancedperipherals.network.MNetwork;
+import de.srendi.advancedperipherals.network.APNetworking;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.Level;
@@ -36,7 +34,6 @@ public class AdvancedPeripherals {
         APConfig.register(ModLoadingContext.get());
 
         modBus.addListener(this::commonSetup);
-        modBus.addListener(this::clientSetup);
         Registration.register();
         MinecraftForge.EVENT_BUS.register(this);
     }
@@ -58,13 +55,9 @@ public class AdvancedPeripherals {
     public void commonSetup(FMLCommonSetupEvent event) {
         APAddons.commonSetup();
         event.enqueueWork(() -> {
-            MNetwork.init();
+            APNetworking.init();
             VillageStructures.init();
         });
-    }
-
-    public void clientSetup(FMLClientSetupEvent event) {
-        HudOverlayHandler.init();
     }
 
 }