+++ src/main/java/de/srendi/advancedperipherals/common/addons/computercraft/owner/TurtleFuelAbility.java
@@ -1,6 +1,5 @@
 package de.srendi.advancedperipherals.common.addons.computercraft.owner;
 
-import dan200.computercraft.ComputerCraft;
 import org.jetbrains.annotations.NotNull;
 
 public class TurtleFuelAbility extends FuelAbility<TurtlePeripheralOwner> {
@@ -23,7 +22,7 @@ public class TurtleFuelAbility extends FuelAbility<TurtlePeripheralOwner> {
 
     @Override
     public boolean isFuelConsumptionDisable() {
-        return !ComputerCraft.turtlesNeedFuel;
+        return !owner.getTurtle().isFuelNeeded();
     }
 
     @Override