+++ src/main/java/de/srendi/advancedperipherals/common/configuration/PeripheralsConfig.java
@@ -16,6 +16,9 @@ public class PeripheralsConfig implements IAPConfig {
     public final ForgeConfigSpec.BooleanValue morePlayerInformation;
     public final ForgeConfigSpec.BooleanValue enablePlayerDetector;
     public final ForgeConfigSpec.BooleanValue playerDetMultiDimensional;
+    public final ForgeConfigSpec.BooleanValue playerSpyRandError;
+    public final ForgeConfigSpec.IntValue playerSpyRandErrorAmount;
+    public final ForgeConfigSpec.IntValue playerSpyPreciseMaxRange;
 
     //Energy Detector
     public final ForgeConfigSpec.IntValue energyDetectorMaxFlow;
@@ -85,6 +88,9 @@ public class PeripheralsConfig implements IAPConfig {
         playerSpy = builder.comment("Activates the \"getPlayerPos\" function of the Player Detector").define("enablePlayerPosFunction", true);
         morePlayerInformation = builder.comment("Adds more information to `getPlayerPos` of the Player Detector. Like rotation and dimension").define("morePlayerInformation", true);
         playerDetMultiDimensional = builder.comment("If true, the player detector can observe players which aren't in the same dimension as the detector itself. `playerDetMaxRange` needs to be infinite(-1) for it to work.").define("chatBoxMultiDimensional", true);
+        playerSpyRandError = builder.comment("If true, add random error to `getPlayerPos` player position that varies based on how far the player is from the detector. Prevents getting the exact position of players far from the detector.").define("enablePlayerPosRandomError", false);
+        playerSpyRandErrorAmount = builder.comment("The maximum amount of error (in blocks) that can be applied to each axis of the player's position.").defineInRange("playerPosRandomErrorAmount", 1000, 0, Integer.MAX_VALUE);
+        playerSpyPreciseMaxRange = builder.comment("If random error is enabled: this is the maximum range at which an exact player position is returned, before random error starts to be applied.").defineInRange("playerPosPreciseMaxRange", 100, 0, Integer.MAX_VALUE);
 
         pop("Energy_Detector", builder);